cicero/backend/requirements.txt
fastapi[all]
uvicorn
pydantic
httpx
weaviate-client
python-dotenv
